+++ contrib/platform/src/com/sun/jna/platform/win32/WinNT.java
@@ -1181,8 +1181,31 @@ public interface WinNT extends WinError, WinDef, WinBase, BaseTSD {
      */
     @FieldOrder({"u"})
     public static class LARGE_INTEGER extends Structure implements Comparable<LARGE_INTEGER> {
+
         public static class ByReference extends LARGE_INTEGER implements
                 Structure.ByReference {
+
+            public ByReference() {
+                super();
+            }
+
+            public ByReference(long value) {
+                super(value);
+            }
+
+        }
+
+        public static class ByValue extends LARGE_INTEGER implements
+                Structure.ByValue {
+
+            public ByValue() {
+                super();
+            }
+
+            public ByValue(long value) {
+                super(value);
+            }
+
         }
 
         @FieldOrder({"LowPart", "HighPart"})